+++ src/test/java/io/github/mmbishop/gwttest/ExceptionTest.java
@@ -16,6 +16,7 @@
 
 package io.github.mmbishop.gwttest;
 
+import io.github.mmbishop.gwttest.core.ExpectedExceptionNotThrownException;
 import io.github.mmbishop.gwttest.core.GwtTest;
 import io.github.mmbishop.gwttest.core.UnexpectedExceptionCaughtException;
 import io.github.mmbishop.gwttest.functions.GwtFunction;
@@ -24,10 +25,18 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 
 public class ExceptionTest {
 
     private final GwtTest<ExceptionContext> gwt = new GwtTest<>(ExceptionContext.class);
+
+    @Test
+    void test_succeeds_when_no_expected_exception_is_declared_and_no_exception_is_thrown() {
+        gwt.test()
+                .when(doing_something_that_does_not_result_in_an_exception)
+                .then(no_exception_is_thrown);
+    }
 
     @Test
     void expected_exception_is_thrown_and_caught() {
@@ -65,15 +74,31 @@ public class ExceptionTest {
             gwt.test()
                     .when(doing_something_that_results_in_an_error)
                     .then(an_exception_is_thrown_and_caught);
-            Assertions.fail();
+            Assertions.fail();  // UnexpectedExceptionCaughtException should have been thrown.
         }
         catch (UnexpectedExceptionCaughtException e) {
-            // Test succeeds if this exception is thrown/
+            // Test succeeds if this exception is thrown.
         }
         catch (Throwable e) {
             Assertions.fail();
         }
     }
+
+    @Test
+    void test_fails_when_expected_exception_is_not_thrown() {
+        try {
+            gwt.test().expectingException(RuntimeException.class)
+                    .when(doing_something_that_does_not_result_in_an_exception);
+            Assertions.fail();  // ExpectedExceptionNotThrownException should have been thrown.
+        }
+        catch (ExpectedExceptionNotThrownException e) {
+            // Test succeeds if this exception is thrown.
+        }
+    }
+
+    private final GwtFunction<ExceptionContext> doing_something_that_does_not_result_in_an_exception = context -> {
+        return;
+    };
 
     private final GwtFunction<ExceptionContext> doing_something_that_results_in_an_exception = context -> {
         throw new RuntimeException("An error occurred while doing something.");
@@ -84,6 +109,10 @@ public class ExceptionTest {
     };
 
     private final GwtFunction<ExceptionContext> an_exception_is_thrown_and_caught = context -> assertNotNull(context.thrownException);
+
+    private final GwtFunction<ExceptionContext> no_exception_is_thrown = context -> {
+        assertNull(context.thrownException);
+    };
 
     public static class ExceptionContext extends Context {}